# use context to cancel compactions