closes https://github.com/prometheus/tsdb/issues/227

Signed-off-by: Krasi Georgiev <kgeorgie@redhat.com>