@krasi-georgiev 2018-12-07
checking the context inside the series loop seems to have a very small impact on the overall compaction time  and if I am reading the numbers right for a block with 20000 series it will increase the time by 0.08sec 

```
func BenchmarkCompactions(t *testing.B) {
	tmpdir, err := ioutil.TempDir("", "test")
	testutil.Ok(t, err)
	defer os.RemoveAll(tmpdir)

	for i := 0; i < t.N; i++ {
		createPopulatedBlock(t, tmpdir, 20000, 0, 1)
	}
}

go test -benchmem -run=^$ github.com/prometheus/tsdb -bench ^BenchmarkCompactions$ -v  -count=1 -benchtime=1m > no-context.txt

go test -benchmem -run=^$ github.com/prometheus/tsdb -bench ^BenchmarkCompactions$ -v  -count=1 -benchtime=1m > with-context.txt

benchcmp no-context.txt with-context.txt
benchmark                  old ns/op     new ns/op     delta
BenchmarkCompactions-8     705443374     784730260     +11.24%

benchmark                  old allocs     new allocs     delta
BenchmarkCompactions-8     832343         832353         +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkCompactions-8     271110513     271112374     +0.00%
```