@krasi-georgiev 2018-12-07
I don't like that I have to add this new metric `populatingBlocks prometheus.Counter` just for the test, but couldn't figure another way to detect that a compaction has  actually started.

@gouthamve any other ideas to avoid this new metric or make it more useful for the actual library usage?