@krasi-georgiev 2019-01-24
the logic is to first check how long it takes the full compaction cycle and than test again when cancelled. This way it should be resistant to CI hiccups, don't you think?
I don't really  like mocking things. If I can't make it pass 10 times in a row or it becomes flaky at some point I will think again. 

> Also where context is used in this scenario (as title suggest). It's all about some channels, there is no context used.. or am I blind?

https://github.com/prometheus/tsdb/pull/465/files#diff-43297cfd6441b52c54c50d646a564339R144