@bwplotka 2019-01-24
Thx for context reference. Looks like a new thing, was looking in older code (: 

> the logic is to first check how long it takes the full compaction cycle and than test again when cancelled which should take less than half the time.  This way it should be resistant to CI hiccups,

No it is not resistant. Anything time related is not resistant.

1. You do full compaction cycle. CI hiccups happens. It took 2m
2. Let's say you have a bug that ignores context cancel. So you do second iteration with interrupt. Because of bug that ignores cancellation, compactor is doing full compaction. There was no CI hiccups so it was faster. It took 5s.

Our test here effectively succeded even though bug was in code. So it gives us 0 confidence. 

So..  Either we do mock even if you don't like this or this test can be deleted. But again, it's only my opinion after some expierience (: 

> If I can't make it pass 10 times in a row or it becomes flaky at some point I will think again.

I can tell you that it will be flaky looking on this code, so I save your time. (: