@krasi-georgiev 2019-02-05
30 re-triggers and not a single failure. Seems good to go. 
If it does happen to cause some issues can always revisit or remove.