@krasi-georgiev 2019-02-06
> > 30 re-triggers and not a single failure. Seems good to go.
> > This does not SGTM. This tests will flake no matter what, and you can test it thousands of times on your fast laptop, but it will flake on slow CI anytime soon. (:

Will buy you a beer every-time it flakes :) if we reach 10 beers It is a good enough motivation for me that I remove the test or find another way :smile: 

I actually run it 30 times over a period of 2 days on Travis itself, but yeah if it ends up causing problems will revisit.