@codesome 2019-02-07
(https://github.com/prometheus/tsdb/pull/465#discussion_r254735655 is not a blocker)