@krasi-georgiev 2019-02-08
> Did you intend to put this sleep outside the `if` statement and inside the `for`?

yep , updated!

> (PS: this loop can be simplified like the above 2 loops which I commented on before)

don't see how can we use the same logic here as it will start a new gorotine on every loop