@codesome 2019-02-08
> don't see how can we use the same logic here as it will start a new gorotine on every loop

```go
for prom_testutil.ToFloat64(db.compactor.(*LeveledCompactor).metrics.populatingBlocks) <= 0 { 
    time.Sleep(3 * time.Millisecond)
}
go func() {
    testutil.Ok(t, db.Close())
    close(dbClosed)
}()
```